make pci functions unsigned

+++ usr/src/uts/i86pc/os/pci_mech1_amd.c
@@ -22,6 +22,7 @@
  * Copyright 2010 Advanced Micro Devices, Inc.
  * Copyright (c) 2010, Oracle and/or its affiliates. All rights reserved.
  * Copyright 2022 Oxide Computer Company
+ * Copyright 2026 RackTop Systems, Inc.
  */
 
 /*
@@ -92,7 +93,7 @@ pci_check_amd_ioecs(void)
  */
 
 uint8_t
-pci_mech1_amd_getb(int bus, int device, int function, int reg)
+pci_mech1_amd_getb(uint8_t bus, uint8_t device, uint8_t function, uint16_t reg)
 {
 	uint8_t val;
 
@@ -113,7 +114,7 @@ pci_mech1_amd_getb(int bus, int device, int function, int reg)
 }
 
 uint16_t
-pci_mech1_amd_getw(int bus, int device, int function, int reg)
+pci_mech1_amd_getw(uint8_t bus, uint8_t device, uint8_t function, uint16_t reg)
 {
 	uint16_t val;
 
@@ -134,7 +135,7 @@ pci_mech1_amd_getw(int bus, int device, int function, int reg)
 }
 
 uint32_t
-pci_mech1_amd_getl(int bus, int device, int function, int reg)
+pci_mech1_amd_getl(uint8_t bus, uint8_t device, uint8_t function, uint16_t reg)
 {
 	uint32_t val;
 
@@ -155,7 +156,8 @@ pci_mech1_amd_getl(int bus, int device, int function, int reg)
 }
 
 void
-pci_mech1_amd_putb(int bus, int device, int function, int reg, uint8_t val)
+pci_mech1_amd_putb(uint8_t bus, uint8_t device, uint8_t function, uint16_t reg,
+    uint8_t val)
 {
 	if (device == PCI_MECH1_SPEC_CYCLE_DEV &&
 	    function == PCI_MECH1_SPEC_CYCLE_FUNC) {
@@ -169,7 +171,8 @@ pci_mech1_amd_putb(int bus, int device, int function, int reg, uint8_t val)
 }
 
 void
-pci_mech1_amd_putw(int bus, int device, int function, int reg, uint16_t val)
+pci_mech1_amd_putw(uint8_t bus, uint8_t device, uint8_t function, uint16_t reg,
+    uint16_t val)
 {
 	if (device == PCI_MECH1_SPEC_CYCLE_DEV &&
 	    function == PCI_MECH1_SPEC_CYCLE_FUNC) {
@@ -183,7 +186,8 @@ pci_mech1_amd_putw(int bus, int device, int function, int reg, uint16_t val)
 }
 
 void
-pci_mech1_amd_putl(int bus, int device, int function, int reg, uint32_t val)
+pci_mech1_amd_putl(uint8_t bus, uint8_t device, uint8_t function, uint16_t reg,
+    uint32_t val)
 {
 	if (device == PCI_MECH1_SPEC_CYCLE_DEV &&
 	    function == PCI_MECH1_SPEC_CYCLE_FUNC) {
